+++ CHANGELOG.md
@@ -10,6 +10,15 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 ### Added
 
 - Add `vrf_seed` to the `/v3/sortitions` rpc endpoint
+
+### Changed
+
+### Fixed
+
+## [3.1.0.0.5]
+
+### Added
+
 - Add miner configuration option `tenure_extend_cost_threshold` to specify the percentage of the tenure budget that must be spent before a time-based tenure extend is attempted
 
 ### Changed